+++ tests/jms-tests/src/test/java/org/apache/activemq/artemis/jms/tests/message/MessageTestBase.java
@@ -194,8 +194,8 @@ public abstract class MessageTestBase extends ActiveMQServerTestCase {
       ProxyAssertSupport.assertNotNull(m);
       ProxyAssertSupport.assertEquals(true, m.getBooleanProperty("booleanProperty"));
       ProxyAssertSupport.assertEquals((byte) 3, m.getByteProperty("byteProperty"));
-      ProxyAssertSupport.assertEquals(Double.valueOf(4.0), Double.valueOf(m.getDoubleProperty("doubleProperty")));
-      ProxyAssertSupport.assertEquals(Float.valueOf(5.0f), Float.valueOf(m.getFloatProperty("floatProperty")));
+      ProxyAssertSupport.assertEquals(4.0, m.getDoubleProperty("doubleProperty"));
+      ProxyAssertSupport.assertEquals(5.0f, m.getFloatProperty("floatProperty"));
       ProxyAssertSupport.assertEquals(6, m.getIntProperty("intProperty"));
       ProxyAssertSupport.assertEquals(7, m.getLongProperty("longProperty"));
       ProxyAssertSupport.assertEquals((short) 8, m.getShortProperty("shortProperty"));
